Split Nuklear theme application
